Crate glyphon

Glyphon provides a simple way to render 2D text with wpgu, cosmic-text and etagere.

Re-exports

pub use cosmic_text;

pub use cosmic_text::fontdb;

Structs

Attrs

Text attributes

AttrsList

List of text attributes to apply to a line

AttrsOwned

An owned version of Attrs

Buffer

A buffer of text that is shaped and laid out

BufferLine

A line (or paragraph) of text that is shaped and laid out

CacheKey

Key for building a glyph cache

Color

Text color

Cursor

Current cursor location

Editor

A wrapper of [Buffer] for easy editing

Font

A font

FontSystem

Access to the system fonts.

LayoutCursor

The position of a cursor within a [Buffer].

LayoutGlyph

A laid out glyph

LayoutLine

A line of laid out glyphs

LayoutRun

A line of visible text for rendering

LayoutRunIter

An iterator of visible text lines, see LayoutRun

Metrics

Metrics of text

Resolution

The screen resolution to use when rendering text.

ShapeGlyph

A shaped glyph

ShapeLine

A shaped line (or paragraph)

ShapeSpan

A shaped span (for bidirectional processing)

ShapeWord

A shaped word (for word wrapping)

SwashCache

Cache for rasterizing with the swash scaler

SwashImage

Scaled glyph image.

TextArea

A text area containing text to be rendered along with its overflow behavior.

TextAtlas

An atlas containing a cache of rasterized glyphs that can be rendered.

TextBounds

Controls the visible area of the text. Any text outside of the visible area will be clipped.

TextRenderer

A text renderer that uses cached glyphs to render text into an existing render pass.

Weight

Specifies the weight of glyphs in the font, their degree of blackness or stroke thickness.

Enums

Action

An action to perform on an Editor

Affinity

Whether to associate cursors placed at a boundary between runs with the run before or after it.

ColorMode

The color mode of an [Atlas].

Command

Path command.

Family

A font family.

FamilyOwned

An owned version of Family

PrepareError

An error that occurred while preparing text for rendering.

RenderError

An error that occurred while rendering text.

Shaping

The shaping strategy of some text.

Stretch

A face width.

Style

Allows italic or oblique faces to be selected.

SubpixelBin

Binning of subpixel position for cache optimization

SwashContent

Content of a scaled glyph image.

Wrap

Wrapping mode

Traits

Edit

A trait to allow easy replacements of Editor, like SyntaxEditor